Implement connecting and connect handler
It seems through testing that the setting connect timeout on the mqtt
library doesn't actually timeout as you might expect. So for now I've
implemented a dodgy timeout pushing an err through the channel if its
still not connected after 5 attempts (1 second breaks)

It also implements the mqtt connect handler by subscribing to the
slack intent from the configuration.

As we can't return any errors we use an error channel to communicate
when an error occurs in the handler and will pass a message through the
channel and exit the program.