Support querying slack users and injecting
Supports the slack token via the configuration file. So we can use that
token to get all the users.

Also implements publishing the entity payload to injections so that it
has a complete list of the support slack names to choose from when doing
the ASR/NLU processing.

The slack refresh is set to every 7 hours on a simple ticker which will
run the entire lifetime of the running process